Dynamically read setup.py meta from package. Fixes #2.
Importing the project module itself requires that `scrapy` be installed
before this package can be installed by easy_install or pip due to it
being a dependency of the main module.

That action breaks common build patterns where `scrapy` &
`scrapy-random-useragent` are installed at the same time. See
the ticket for more information.

The fix here is to simply open the package file in setup.py and take the
variables we need directly out of it; this solution means we can keep
things DRY and don't have to repeat meta & version information across
the project.